update config
removed and organized plugins. readme update
