+++ README.md
@@ -1,13 +1,7 @@
-## TODO
+# TODO
 
-### Fix
+## Fix
 
-- **LSP**
-  - [x] Dart/Flutter
-  - [x] Lua
-  - [x] Go
-  - [x] Vue.js
-  - [x] React.js
-  - [x] Autocompletion(lua)
-  - [x] Formatting
-- removed dap
+- [ ] Go
+- [ ] Vue.js
+- [ ] React.js
