--- lua/plugins/autoclose.lua
@@ -1,29 +1,0 @@
-return {
-  {
-    "m4xshen/autoclose.nvim",
-    config = function()
-      require("autoclose").setup()
-    end,
-  },
-  {
-    "windwp/nvim-ts-autotag",
-    ft = {
-      "html",
-      "javascript",
-      "javascriptreact",
-      "typescript",
-      "typescriptreact",
-      "vue",
-      "xml"
-    },
-    config = function()
-      require("nvim-ts-autotag").setup({
-        opts = {
-          enable_close = true,          -- Auto close tags
-          enable_rename = true,         -- Auto rename pairs of tags
-          enable_close_on_slash = false -- Auto close on trailing </
-        }
-      })
-    end,
-  }
-}
