+++ lua/plugins/better-escape.lua
@@ -1,6 +1,7 @@
 -- lua with lazy.nvim
 return {
   "max397574/better-escape.nvim",
+  enabled = false,
   config = function()
     require("better_escape").setup {
       event = "VeryLazy",
