--- lua/plugins/colorizer.lua
@@ -1,8 +1,0 @@
-return {
-  {
-    'norcalli/nvim-colorizer.lua',
-    config = function()
-      require("colorizer").setup()
-    end
-  }
-}
