--- lua/plugins/hardtime.lua
@@ -1,8 +1,0 @@
-return {
-  {
-    "m4xshen/hardtime.nvim",
-    dependencies = { "MunifTanjim/nui.nvim", "nvim-lua/plenary.nvim" },
-    opts = {},
-    enabled = false,
-  }
-}
