--- lua/plugins/indent-blackline.lua
@@ -1,28 +1,0 @@
-return {
-  {
-    "lukas-reineke/indent-blankline.nvim",
-    main = "ibl",
-    opts = {},
-    config = function()
-      require("ibl").setup({
-        indent = {
-          char = "│", -- Use a thin vertical line for indents
-        },
-        scope = {
-          enabled = true,
-          show_start = false,
-          show_end = false,
-        },
-        exclude = {
-          filetypes = {
-            "help",
-            "terminal",
-            "lspinfo",
-            "TelescopePrompt",
-            "TelescopeResults",
-          },
-        },
-      })
-    end,
-  },
-}
