--- lua/plugins/resession.lua
@@ -1,8 +1,0 @@
-return {
-  {
-    'stevearc/resession.nvim',
-    opts = {},
-    config = function()
-    end
-  }
-}
