--- lua/plugins/todo-comment.lua
@@ -1,48 +1,0 @@
-return {
-  {
-    "folke/todo-comments.nvim",
-    dependencies = { "nvim-lua/plenary.nvim" },
-    opts = {
-      -- your configuration comes here
-      -- or leave it empty to use the default settings
-      -- refer to the configuration section below
-    }
-  },
-  {
-    "folke/trouble.nvim",
-    opts = {}, -- for default options, refer to the configuration section for custom setup.
-    cmd = "Trouble",
-    -- keys = {
-    --   {
-    --     "<leader>xx",
-    --     "<cmd>Trouble diagnostics toggle<cr>",
-    --     desc = "Diagnostics (Trouble)",
-    --   },
-    --   {
-    --     "<leader>xX",
-    --     "<cmd>Trouble diagnostics toggle filter.buf=0<cr>",
-    --     desc = "Buffer Diagnostics (Trouble)",
-    --   },
-    --   {
-    --     "<leader>xs",
-    --     "<cmd>Trouble symbols toggle focus=false<cr>",
-    --     desc = "Symbols (Trouble)",
-    --   },
-    --   {
-    --     "<leader>xl",
-    --     "<cmd>Trouble lsp toggle focus=false win.position=right<cr>",
-    --     desc = "LSP Definitions / references / ... (Trouble)",
-    --   },
-    --   {
-    --     "<leader>xL",
-    --     "<cmd>Trouble loclist toggle<cr>",
-    --     desc = "Location List (Trouble)",
-    --   },
-    --   {
-    --     "<leader>xQ",
-    --     "<cmd>Trouble qflist toggle<cr>",
-    --     desc = "Quickfix List (Trouble)",
-    --   },
-    -- },
-  },
-}
